refactor(html): Add data attributes to game board
This commit adds `data-row` and `data-col` attributes to the
HTML elements representing the Tic-Tac-Toe board tiles

These data attributes store the row and column indices for each tile,
which will be used by the DOMView to identify the clicked tile and
update the game state. This allows the DOMView to easily determine
which cell the user clicked.